heritage_graph/requirements.txt
asgiref==3.8.1
attrs==25.3.0
begins==0.9
black==25.1.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.2.1
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.6.0
django-environ==0.12.0
django-erd==1.0
django-extensions==3.2.3
django-filter==25.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dot==0.3.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
ecdsa==0.19.1
eralchemy==1.5.0
filelock==3.19.1
funcsigs==1.0.2
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
identify==2.6.13
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.3.0
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pycparser==2.22
pydot==3.0.2
PyJWT==2.10.0
pyparsing==3.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.12.9
setuptools==78.1.1
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uv==0.8.0
virtualenv==20.34.0
wheel==0.45.1
whitenoise==6.9.0
